+++ lib/feature_discovery.dart
@@ -1,4 +1,12 @@
 library feature_discovery;
 
-export 'src/feature_discovery.dart';
-export 'src/ensure_visible.dart';
+// The library files (foundation.dart, widgets.dart, etc.) are in src
+// because they should not appear for code completion because the
+// members that should be exposed when using the package are exposed here.
+export 'src/widgets.dart'
+    show
+        DescribedFeatureOverlay,
+        OverflowMode,
+        EnsureVisible,
+        EnsureVisibleState;
+export 'src/foundation.dart' show FeatureDiscovery, ContentOrientation;
