+++ lib/src/foundation.dart
@@ -1,0 +1,2 @@
+export 'foundation/feature_discovery.dart';
+export 'foundation/bloc.dart';
