+++ lib/src/foundation/bloc.dart
